refactor: Refactor Review entity
- ReviewReport와 ReviewReaction이 있으면 Review가 삭제되지 않는 경우 해결하기 위해 수정
- 파일 위치 변경
