+++ src/main/java/com/personal/contentservice/domain/Review.java
@@ -1,12 +1,12 @@
 package com.personal.contentservice.domain;
 
-import jakarta.persistence.Column;
+import jakarta.persistence.CascadeType;import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinColumns;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
+import jakarta.persistence.OneToMany;import jakarta.persistence.Table;
+import java.util.List;import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -44,5 +44,11 @@ public class Review extends BaseEntity {
 
   @Column(nullable = false)
   private int reportedCount;
+
+  @OneToMany(mappedBy = "review", cascade = CascadeType.ALL, orphanRemoval = true)
+  private List<ReviewReaction> reactions;
+
+  @OneToMany(mappedBy = "review", cascade = CascadeType.ALL, orphanRemoval = true)
+  private List<ReviewReport> reports;
 
 }
