+++ src/main/java/com/personal/contentservice/domain/User.java
@@ -30,9 +30,6 @@ public class User extends BaseEntity {
   private String password;
 
   @Enumerated(EnumType.STRING)
-  private UserType userType;
-
-  @Enumerated(EnumType.STRING)
   private UserStatus userStatus;
 
 }
