+++ src/main/java/com/personal/contentservice/dto/review/ReviewReactionDto.java
@@ -1,5 +1,6 @@
-package com.personal.contentservice.dto;
+package com.personal.contentservice.dto.review;
 
+import com.personal.contentservice.type.ReactionType;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,12 +11,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReviewReportDto {
-
-  @NotNull
-  private String reason;
+public class ReviewReactionDto {
 
   @NotNull
   private long reviewId;
+
+  @NotNull
+  private ReactionType reactionType;
 
 }
