+++ src/main/java/com/personal/contentservice/service/impl/ReviewServiceImpl.java
@@ -11,8 +11,8 @@ import com.personal.contentservice.domain.Review;
 import com.personal.contentservice.domain.ReviewReaction;
 import com.personal.contentservice.domain.ReviewReport;
 import com.personal.contentservice.domain.User;
-import com.personal.contentservice.dto.ReviewReactionDto;
-import com.personal.contentservice.dto.ReviewReportDto;
+import com.personal.contentservice.dto.review.ReviewReactionDto;
+import com.personal.contentservice.dto.review.ReviewReportDto;
 import com.personal.contentservice.dto.review.ReviewAddDto;
 import com.personal.contentservice.dto.review.ReviewDeleteDto;
 import com.personal.contentservice.dto.review.ReviewUpdateDto;
