+++ src/main/resources/application.yml
@@ -1,34 +1,24 @@
 spring:
   datasource:
-    url: jdbc:h2:mem:content;NON_KEYWORDS=USER
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost:3306/content?serverTimezone=UTC&characterEncoding=UTF-8
     username: content
-    password:
-    driverClassName: org.h2.Driver
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
+    password: content
+
   jpa:
-    defer-datasource-initialization: true
-    database-platform: H2
     hibernate:
-      ddl-auto: create-drop
-    open-in-view: false
+      ddl-auto: update
+    generate-ddl: true
+
     properties:
+      format_sql: true
       hibernate:
-        format_sql: true
-        show_sql: true
+        show-sql: true
 
   data:
     redis:
       host: localhost
       port: 6379
-    elasticsearch:
-      host: 127.0.0.1
-      port: 9200
-      indices:
-        contentsIndexName: contents
-        testIndexName: tests
 
 tmdb:
   api:
