+++ src/test/java/com/personal/contentservice/service/impl/ReviewServiceImplTest.java
@@ -13,8 +13,8 @@ import com.personal.contentservice.domain.Content;
 import com.personal.contentservice.domain.ContentKey;
 import com.personal.contentservice.domain.Review;
 import com.personal.contentservice.domain.User;
-import com.personal.contentservice.dto.ReviewReactionDto;
-import com.personal.contentservice.dto.ReviewReportDto;
+import com.personal.contentservice.dto.review.ReviewReactionDto;
+import com.personal.contentservice.dto.review.ReviewReportDto;
 import com.personal.contentservice.dto.review.ReviewAddDto;
 import com.personal.contentservice.dto.review.ReviewDeleteDto;
 import com.personal.contentservice.dto.review.ReviewUpdateDto;
